requirements.txt
# Core Bot Framework
python-telegram-bot==20.7

# Image Rendering (statistics cards)
playwright==1.40.0

# Image Processing
Pillow==10.1.0
pyzbar==0.1.9

# Environment Variables
python-dotenv==1.0.0

# QR Code
qrcode==7.4.2

# Utilities
python-dateutil==2.8.2
